requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Simulation
simpy>=4.0.1

# Probabilistic modeling
pymc>=5.0.0
arviz>=0.15.0
scipy>=1.10.0

# Machine Learning & Optimization
scikit-learn>=1.3.0
networkx>=3.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data handling
openpyxl>=3.1.0
xlrd>=2.0.1

# Configuration & Logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
jupyter>=1.0.0
ipython>=8.12.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced probabilistic programming
tensorflow-probability>=0.20.0
